Applied rsc structure to mobile rsc sample
Moved the NavigationStack back into App because for readability. Only need to move it out if configure function props - would rather explain that in the docs than have that as the default in the sample
