+++ NavigationReactMobile/sample/rsc/src/App.tsx
@@ -1,7 +1,8 @@
 'use server-entry';
 import './client';
-import { Scene } from 'navigation-react-mobile';
-import RootProvider from './RootProvider';
+import { NavigationStack, Scene } from 'navigation-react-mobile';
+import NavigationProvider from './NavigationProvider';
+import HmrProvider from './HmrProvider';
 import People from './People';
 import Person from './Person';
 
@@ -12,14 +13,21 @@ const App = async ({url}: any) => {
         <title>Navigation React Mobile</title>
       </head>
       <body>
-        <RootProvider url={url}>
-          <Scene stateKey="people">
-            <People />
-          </Scene>
-          <Scene stateKey="person" dataKeyDeps={['id']}>
-            <Person />
-          </Scene>
-        </RootProvider>
+        <NavigationProvider url={url}>
+          <HmrProvider>
+            <NavigationStack
+              unmountStyle={[{transform: 'translateX(100%)'}, {transform: 'translateX(0)'}]}
+              crumbStyle={[{transform: 'translateX(5%) scale(0.9)', opacity: 0},{transform: 'translateX(0) scale(1)', opacity: 1}]}
+              style={{position: 'fixed', left: '0', right: '0', top: '0', bottom: '0', overflow: 'auto', backgroundColor: '#fff', margin: '8px'}}>
+                <Scene stateKey="people">
+                  <People />
+                </Scene>
+                <Scene stateKey="person" dataKeyDeps={['id']}>
+                  <Person />
+                </Scene>
+            </NavigationStack>
+          </HmrProvider>
+        </NavigationProvider>
       </body>
     </html>
   );
