--- NavigationReactMobile/sample/rsc/src/HmrContext.ts
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export default createContext<any>(null);
